[3/4] sqlite query perf: clean up in-flight statements on cursor close
(...or deactivate, even though it's deprecated)

Let the SQLiteQuery hold a WeakReference<PreparedStatement> so that we
can tell the SQLiteConnection when there's no longer a use for avoiding
resetting that statement.

Since PreparedStatements can be pooled or reused, we also have to keep
a WeakReference to the "last client" (i.e. the SQLiteQuery) from the
statement, so that we can check that it's still the right one.

If the connection is available (not acquired by a thread), the reset
will happen immediately. Otherwise, it be kept in a queue on the
connection, and the reset will happen when the connection is released.
To avoid races, the queueing and clearing is always done while holding
SQLiteConnectionPool.mLock.

Change-Id: I30fe9a72ef7c08cc5adcf233cca2ae216b97eb86
Signed-off-by: Corey Edwards <ensabahnur16@gmail.com>