Skip duplicated appop checking of ContentProvider.
If provider has set required op by setAppOps, and the op
is the same as the one associated with permission, it is
no need to check twice.

Also make permissionToOp to be consistent because there
may have multiple permissions map to one op. For the case,
give it switch op rather than the last op. (e.g. READ_SMS)

Change-Id: I08bddbe66655c19ffa0ba13a26163fd1b1385741