Access mView in the thread which created the ViewRootImpl
Accessing ViewRootImpl.mView should be only in the single thread which
creates the ViewRootImpl object. Usually, it is the main thread of the
process. However, ViewRootImpl is sometimes not created in the main
thread, which causes there would be two threads accessing
ViewRootImpl.mView. One is the thread which creates ViewRootImpl, the
other one is the main thread which would call
AccessibilityStateChangeListener.onAccessibilityStateChanged().

In this patch, we post a message to dispatch the accessibility event
in the thread of ViewRootImpl, and make sure ViewRootImpl.mView is not
null before accessing it.

Change-Id: I20eec407894d9ebba9d1c12cf31a3210ac34b85f